requirements.txt
torch
torchvision
segmentation-models-pytorch
ultralytics
opencv-python
numpy
requests
Pillow
shapely
dotenv
shapely